Shtepia/Document
Rifaktorimi i kodit per Shtepia Controller dhe implementimi i logjikes per gjenerimin e dokumentit pas qe krijohet nje shtepi.
