+++ Real-Estate-App/Application/Features/Prona/DocumentFeature.cs
@@ -3,6 +3,7 @@ using Domain.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection.Metadata;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -23,6 +24,22 @@ namespace Application.Features.Prona
             {
                 Type = "Apartment Document",
                 PronaID = apartmentId,
+                CreatedData = DateTime.Now,
+                ExpiorationDate = DateTime.Now.AddYears(1)
+            };
+
+            _context.Documents.Add(document);
+            await _context.SaveChangesAsync();
+
+            return document;
+        }
+
+        public async Task<Documents> CreateDocumentForShtepiaAsync(int shtepiaId)
+        {
+            var document = new Documents
+            {
+                Type = "Shtepia Document",
+                PronaID = shtepiaId,
                 CreatedData = DateTime.Now,
                 ExpiorationDate = DateTime.Now.AddYears(1)
             };
