+++ Real-Estate-App/Application/Features/Prona/ShtepiaFeature.cs
@@ -1,0 +1,62 @@
+﻿using Application.DTO;
+using Application.Interface;
+using Domain.Entities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Application.Features.Prona
+{
+    public class ShtepiaFeature
+    {
+        private readonly IAppDbContext _context;
+
+        public ShtepiaFeature(IAppDbContext context)
+        {
+            _context = context;
+        }
+
+        public async Task<Shtepia> CreateShtepiaAsync(ShtepiaCreateDto shtepiaDto)
+        {
+            string photoPath = null;
+
+            if (shtepiaDto.Photo != null)
+            {
+                var uploadsFolder = Path.Combine(Directory.GetCurrentDirectory(), "Images", "Shtepia-Img");
+                Directory.CreateDirectory(uploadsFolder);
+
+                var fileName = $"{Guid.NewGuid()}{Path.GetExtension(shtepiaDto.Photo.FileName)}";
+                var filePath = Path.Combine(uploadsFolder, fileName);
+
+                using (var fileStream = new FileStream(filePath, FileMode.Create))
+                {
+                    await shtepiaDto.Photo.CopyToAsync(fileStream);
+                }
+
+                photoPath = Path.Combine("Images", "Shtepia-Img", fileName);
+            }
+
+            var shtepia = new Shtepia
+            {
+                Emri = shtepiaDto.Emri,
+                Adresa = shtepiaDto.Adresa,
+                Price = shtepiaDto.Price,
+                Description = shtepiaDto.Description,
+                Status = shtepiaDto.Status,
+                Photo = photoPath,
+                size = shtepiaDto.size,
+                nrFloors = shtepiaDto.nrFloors,
+                kaGarazhd = shtepiaDto.kaGarazhd,
+                kaPool = shtepiaDto.kaPool
+            };
+
+            _context.Shtepiat.Add(shtepia);
+            await _context.SaveChangesAsync();
+
+            return shtepia;
+        }
+    }
+
+}
